move files into v1 keypad dir
