+++ packages/math-input/src/components/keypad-v1/iconography/cube-root.js
