+++ packages/math-input/src/components/keypad-v1/iconography/geq.js
