+++ packages/math-input/src/components/keypad-v1/iconography/jump-out-numerator.js
