+++ packages/math-input/src/components/keypad-v1/iconography/ln.js
