+++ packages/math-input/src/components/keypad-v1/iconography/plus.js
